mainte

+++ database/migrations/2023_06_25_101138_create_client_corporations_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
             $table->string('clientcorporation_name')->comment('法人名称');
             $table->string('clientcorporation_kana_name')->comment('法人カナ名称');
             $table->string('clientcorporation_short_name')->comment('法人略称');
-            $table->unsignedInteger('credit_limit')->default(0)->comment('与信限度額');
+            $table->decimal('credit_limit',10, 0)->default(0)->comment('与信限度額');
             $table->text('memo')->nullable(true)->commebt('法人メモ');
             $table->foreignId('created_by')->nullable(true)->comment('作成者');
             $table->foreignId('updated_by')->nullable(true)->comment('更新者');
